renaming useHomeGripperZAxis
